Getting ready to run in production

+++ test/testbase.py
@@ -88,6 +88,9 @@ class TestBaseSaveOutput(abc.ABC, unittest.TestCase):
 
 class TestBaseCopyInput(abc.ABC, unittest.TestCase):
   @classmethod
+  def removecopiedinput(cls): return True
+
+  @classmethod
   @abc.abstractmethod
   def filestocopy(cls): pass
 
@@ -101,5 +104,6 @@ class TestBaseCopyInput(abc.ABC, unittest.TestCase):
   @classmethod
   def tearDownClass(cls):
     super().tearDownClass()
-    for copyfrom, copytofolder in cls.filestocopy():
-      (copytofolder/copyfrom.name).unlink()
+    if cls.removecopiedinput():
+      for copyfrom, copytofolder in cls.filestocopy():
+        (copytofolder/copyfrom.name).unlink()
